code/requirements.txt
# Core
anthropic>=0.7.0
python-dotenv>=0.19.0
click>=8.0.0
requests>=2.28.0

# HTTP Server
flask>=2.0.0

# GitHub Integration
PyGithub>=1.57

# Notion Integration
notion-client>=0.7.0

# Google Drive Integration
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.50.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
python-dev-tools>=2021.2.0
